requirements.txt
# Core dependencies
torch>=1.11.0
torchvision>=0.12.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
# Use faiss-gpu>=1.7.0 if you want GPU acceleration for FAISS

# Data processing
numpy>=1.21.0
Pillow>=9.0.0
datasets>=2.0.0

# UI and visualization
gradio>=4.0.0

# Utilities
tqdm>=4.64.0
requests>=2.28.0